+++ components/SearchInput.js
@@ -2,6 +2,7 @@ import { SearchIcon } from '@heroicons/react/outline';
 import { XCircleIcon } from '@heroicons/react/solid';
 import { bool, func, string } from 'prop-types';
 import { useEffect, useRef, useState } from 'react';
+import useDebounce from '~/hooks/useDebounce';
 import styles from '~/styles/components/SearchInput.module.css';
 
 export default function SearchInput({
@@ -14,16 +15,20 @@ export default function SearchInput({
   const [searchTerm, setSearchTerm] = useState('');
   const inputElement = useRef(null);
 
+  const debouncedSearchTerm = useDebounce(searchTerm, 500);
+
   useEffect(() => {
     if (autoFocus) {
       inputElement.current?.focus();
     }
   }, []);
 
-  function submit(ev) {
-    ev.preventDefault();
-    onFormSubmit({ term: searchTerm });
-  }
+
+  useEffect(() => {
+    if (debouncedSearchTerm) {
+      onFormSubmit({ term: searchTerm });
+    }
+  }, [debouncedSearchTerm]);
 
   function clear() {
     setSearchTerm('');
@@ -31,7 +36,10 @@ export default function SearchInput({
   }
 
   return (
-    <form className={styles.form} onSubmit={submit}>
+    <form
+      className={`${styles.form} search-input rel`}
+      onSubmit={(ev) => ev.preventDefault()}
+    >
       <label className={styles.label}>
         <div className={styles.magnifier}>
           <SearchIcon width={24} height={24} />